Revert "♻️ refactor: CurrencyList 화면 구성 MVC 구조로 분리"
